Setup the user creation endpoint
