+++ src/models/user.js
@@ -1,0 +1,47 @@
+const mongoose = require('mongoose')
+const validator = require('validator')
+
+
+const User = mongoose.model('User',{
+    name: {
+        type: String,
+        required: true,
+        trim: true
+    },
+    email:{
+      type: String,
+      required:true,
+      trim:true,
+      lowercase: true,
+      validate(value) {
+          if(!validator.isEmail(value)){
+              throw new Error('The email is invalid')
+          }
+      }
+    
+    },
+    
+    age:{
+        type: Number,
+        default: 0,
+        validate(value){
+          if (value < 0){
+          throw new Error('Age must be positive')
+          }
+        }
+    },
+    password:{
+        type:String,
+        required:true,
+        trim:true,
+        minlength:7,
+        validate(value) {
+            if(value.toLowerCase().includes("password")){
+                throw new Error('it must not include password')
+            }
+        }
+    
+    }
+    })
+
+    module.exports = User
